A TON OF STUFF WAS WRONG
The avr state machine was clipping the last byte.
The genavr was not making the msg functions virtual.
The avr state machien could allow the buffer to overflow and kill the program
The ros_string did not correclty deal with a max string length
	(I.E., once the max string length is set, it cannot be changed)

There is now a packet time limit.  If a packet takes longer than 30 ms
the state machine resets.  This is to protect against lost transmissions.